Extract bump-dependencies from index :scissors:
